feat(ui): TE-1211 moved alert copy function into utils file
